Remove webpack-dev-server and move test_api to api
Get rid of the original 'webpack-dev-server' environment.
It has been fully replaced and works better than before.

What to do with the api? Depends how rigorous the coupling
will be with the actual api (python development). 'api'
may be more confusing than 'test_api' if a simple relay
is added. However, if a 'rigorous' mapping schema is
added, it may ensure less breakage, but at the cost
of dev insanity. To be continued ...