Adjust reducers to import multiple from a file and add schemas/selectors
I think the goal should be to eliminate any occurrences of `byId`
(except in tests). This means adding selectors to wrap this selection,
instead of calling `state.byId[someId]` in `mapStateToProps`. This
is already covered by the reducers' boilerplate with the standard
format of `export const someReducer = (state, action) => someById(state.byId,
action`)